feat: cleanGMKeys
