+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.3.7
 
 - 新增：评论区 修复笔记缩略图比例
+- 新增：清理插件存储中的弃用项
 - 修复：首页 隐藏全部骨架影响轮播图
 - 优化：视频过滤与 greasyfork 481629 的兼容性
 - 更新：部分功能细节
