+++ src/modules/filters/variety/dynamic/pages/space.ts
@@ -206,6 +206,7 @@ export const dynamicFilterSpaceGroups: Group[] = [
                 type: 'number',
                 id: GM_KEYS.black.duration.valueKey,
                 name: '设定最低时长（0~300s）',
+                noStyle: true,
                 minValue: 0,
                 maxValue: 300,
                 step: 1,
